Bugfix and a little change for the classifier
Now the classifier returns UNKNOWN when countUncertain + the lowest counter isn't less than the highest counter, e.g. with K = 10 if we have countUncertain = 2 and countClassifyFalse = 3 we have countClassifyTrue = 5 which isn't greater than the half so it would make sense to return UNKNOWN rather than feasible which was returned in the old version.
Other than that, now using the linear transform also when the number of possible voting isn't greater than the number of possible queues (there was a problem in the modulo calculation, and I can just use the linear transform so why not) and removed the rankingArray variable since I can directly use this.queueRanking and the other change is that if the number of possible voting values is 1 I can just set the index infeasibility directly like when the number of possible queues is 1, since I have no other choices