Minor changes
There's no need to distinguish between feasible and infeasible items when there's a specific core relation with the query.

I'm now doing other tests: by adding a setter (for the array A) to the ManyInfeasiblePC example class, I can see Evosuite fails less frequently (when generating a test case for satisfiable path conditions) than before and so the path coverage seems to be higher on average.